requirements.txt
pytest-asyncio>=1.0.0
pytest-cov>=6.1.1
pytest-mock>=3.14.1
click==8.1.7
PyYAML==6.0.1
boto3==1.38.33
openai==1.16.1
azure-identity==1.15.0
azure-mgmt-resource==23.1.0
azure-mgmt-monitor>=5.0.0
google-cloud-storage==3.1.0
google-cloud-compute==1.10.0
google-generativeai==0.5.0
python-dotenv>=1.0.0

